Validation

+++ Backend/routes.js
@@ -1,9 +1,20 @@
 const express = require("express");
 const schema = require("./schema");
 const router = express.Router();
+const mongoose = require("mongoose");
+const validationSchema = require("./uservalidation");
 
 const app = express();
 app.use(express.json());
+
+const validateRequest = (req, res, next) => {
+  const { error } = validationSchema.validate(req.body, { abortEarly: false });
+  if (error) {
+    const errorMessages = error.details.map((detail) => detail.message);
+    return res.status(400).json({ error: errorMessages });
+  }
+  next();
+};
 
 router.get("/", async (req, res) => {
   try {
@@ -32,7 +43,7 @@ router.get("/:id", async (req, res) => {
   }
 });
 
-router.post("/", async (req, res) => {
+router.post("/", validateRequest, async (req, res) => {
   try {
     // console.log(req.body);    //To check what is getting posted
     const newSuperstition = await schema.create(req.body);
@@ -47,10 +58,15 @@ router.post("/", async (req, res) => {
   }
 });
 
-router.patch("/:id", async (req, res) => {
+router.patch("/:id", validateRequest, async (req, res) => {
   try {
     const _id = req.params.id;
-    const getSuperstition = await schema.findByIdAndUpdate(_id, req.body);
+    if (!mongoose.Types.ObjectId.isValid(_id)) {
+      return res.status(400).send("Invalid ObjectId");
+    }
+    const getSuperstition = await schema.findByIdAndUpdate(_id, req.body, {
+      new: true,
+    });
     if (getSuperstition) {
       res.status(200).send(getSuperstition);
     } else {
@@ -64,6 +80,9 @@ router.patch("/:id", async (req, res) => {
 router.delete("/:id", async (req, res) => {
   try {
     const _id = req.params.id;
+    if (!mongoose.Types.ObjectId.isValid(_id)) {
+      return res.status(400).send("Invalid ObjectId");
+    }
     const getSuperstition = await schema.findByIdAndDelete(_id);
     if (getSuperstition) {
       res.status(200).send(getSuperstition);
